Initial code commit
The code includes lots of functions pulled over from plot-attack-graph.py which need to be incorporated into the new classes or deleted. some of it won't work.
Not clear yet if the algorithm itself actually works. It is unoptimised and so takes a long time on anything other than a tiny system model.