Shows intermediate cause long description rather than the short one
This is useful because sometimes the short ones are the same even though the long ones are different and it looked as though the same line in the CSV output was being printed twice for no reason.

Also added __eq__ and hash methods to the Entity while debugging this: they aren't strictly needed as it didn't change the behaviour but good to be explicitt perhaps.